+++ src/components/depot/implementation/ipfs-default-pkg.ts
@@ -1,13 +1,19 @@
 import * as Ipfs from "./ipfs/index.js"
 import * as IpfsBase from "./ipfs.js"
+import { Dependencies } from "./ipfs/node.js"
 import { Implementation } from "../implementation.js"
 
 
 // 🛳
 
 
-export async function implementation(peersUrl: string, repoName: string): Promise<Implementation> {
+export async function implementation(
+  dependencies: Dependencies,
+  peersUrl: string,
+  repoName: string
+): Promise<Implementation> {
   const [ ipfs, repo ] = await Ipfs.nodeWithPkg(
+    dependencies,
     await Ipfs.pkgFromCDN(Ipfs.DEFAULT_CDN_URL),
     peersUrl,
     repoName,