+++ tests/fs/exchange.node.test.ts
@@ -1,13 +1,16 @@
 import expect from "expect"
+import { crypto } from "../helpers/components.js"
 import { emptyFilesystem } from "../helpers/filesystem.js"
+
+import { addPublicExchangeKey, hasPublicExchangeKey } from "../../src/fs/data.js"
 
 
 describe("the filesystem", () => {
 
   it("adds public exchange key to well-known location", async function () {
     const fs = await emptyFilesystem()
-    await fs.addPublicExchangeKey()
-    const exists = await fs.hasPublicExchangeKey()
+    await addPublicExchangeKey(crypto, fs)
+    const exists = await hasPublicExchangeKey(crypto, fs)
 
     expect(exists).toEqual(true)
   })